# openapi.yaml
openapi: 3.0.0
info:
  title: API de Mensajes
  description: API para guardar, recuperar y eliminar mensajes de texto en MongoDB.
  version: 1.0.0
servers:
  - url: https://couto.onrender.com

paths:
  /message:
    post:
      summary: Guardar un nuevo mensaje de texto
      operationId: guardarMensaje
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - keyword
                - nombre
                - contenido
              properties:
                keyword:
                  type: string
                  description: Clave de autorización para guardar el mensaje
                nombre:
                  type: string
                  description: Identificador único del mensaje
                contenido:
                  type: string
                  description: Contenido del mensaje en texto plano
      responses:
        '201':
          description: Mensaje guardado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                  new_message:
                    type: object
                    additionalProperties: true
        '403':
          description: Clave incorrecta
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '400':
          description: Error en la solicitud
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    get:
      summary: Recuperar un mensaje por nombre
      operationId: obtenerMensaje
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
          description: Clave de autorización para obtener el mensaje
        - name: nombre
          in: query
          required: true
          schema:
            type: string
          description: Identificador único del mensaje
      responses:
        '200':
          description: Mensaje encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  contenido:
                    type: string
        '403':
          description: Clave incorrecta
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: Mensaje no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
        '400':
          description: Error en la solicitud
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

    delete:
      summary: Eliminar un mensaje por nombre
      operationId: eliminarMensaje
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
          description: Clave de autorización para eliminar el mensaje
        - name: nombre
          in: query
          required: true
          schema:
            type: string
          description: Identificador único del mensaje
      responses:
        '200':
          description: Mensaje eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
        '403':
          description: Clave incorrecta
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
        '404':
          description: Mensaje no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string

  /messages:
    delete:
      summary: Eliminar todos los mensajes (requiere clave)
      operationId: eliminarTodos
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
          description: Clave de autorización para eliminar todos los mensajes
      responses:
        '200':
          description: Todos los mensajes eliminados
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
        '403':
          description: Clave incorrecta
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string